adicionando métodos nos controllers das telas
